requirements.txt
Flask==2.1.1
RPi.GPIO==0.7.0